+++ CHANGES.md
@@ -7,10 +7,15 @@ Beta Releases
 ### b17 - 2013-06-03
 
 * Breaking changes:
+   * Replaced tessellators and meshes with geometry.  In particular:
+      * Replaced `CubeMapEllipsoidTessellator` with `EllipsoidGeometry`.
+      * Renamed `MeshFilters` to `GeometryFilters`.
+      * Renamed `MeshFilters.toWireframeInPlace` to `GeometryFilters.toWireframe`.
    * Renamed `ComponentDatatype.*.toTypedArray` to `ComponentDatatype.*.createTypedArray`.
    * Replaced `Uniform.getFrameNumber` and `Uniform.getTime` with `Uniform.getFrameState`, which returns the full frame state.    
 * Improved the performance of drawing polygons created with `configureFromPolygonHierarchy`.
-* Added `MeshFilters.combine` to combine meshes for better batching.
+* Added `GeometryFilters.combine` to combine meshes for better batching.
+* Added `BoundingSphere.fromEllipsoid`.
 * Added `DrawCommand.cull` to avoid redundant visibility checks.
 * Added `czm_morphTime` automatic GLSL uniform.
 * Added support for floating-point textures.